Rebotes y latidoen funcionamiento
*Se modifica el TIMER0 para obtener el tiempo real.
*Se implementa el latido a 2 Hz
*Se modifica ampliamente la gestión de rebotes para tener en cuenta las nuevas características
*Se modifica el 8led para mostrar el número 6 correctamente
*Se arregla la gestión de la pila de depuración
*El incremento/decremento del 8led funciona correctamente con los botones.

+++ Othello/pilaDebug.h
@@ -2,8 +2,6 @@
 #define _PILADEBUG_H_
 #include "stdint.h"
 
-
-void push_debug(uint32_t ID_evento, uint32_t auxData);
-
+void push_debug(uint8_t ID_evento, uint32_t auxData);
 
 #endif
